requirements.txt
Pillow==7.1.1
pkg-resources==0.0.0
xlrd==1.2.0
